requirements.txt
bcrypt==4.0.1
cryptography==41.0.5
mysql-connector-python==8.2.0
pyAesCrypt==6.1.1
pyOpenSSL==23.3.0
SQLAlchemy==2.0.23
PyNaCl==1.5.0
python-dotenv==1.0.0
customtkinter==5.2.1
requests==2.25.1
urllib3==1.26.5
packaging
pycryptodome==3.19.0
pillow